Fix isXJoinResource

+++ controllers/iteration.go
@@ -168,6 +168,24 @@ func (i *ReconcileIteration) removeFinalizer() error {
 	return i.Client.Update(ctx, i.Instance)
 }
 
+func (i *ReconcileIteration) isXJoinResource(resourceName string) bool {
+	var response bool
+
+	if strings.Index(resourceName, "xjoin-connect-config") == 0 {
+		response = false
+	} else if strings.Index(resourceName, "xjoin-connect-offsets") == 0 {
+		response = false
+	} else if strings.Index(resourceName, "xjoin-connect-status") == 0 {
+		response = false
+	} else if strings.Index(resourceName, i.parameters.ResourceNamePrefix.String()) == 0 {
+		response = true
+	} else if strings.Index(resourceName, database.ReplicationSlotPrefix(i.parameters.ResourceNamePrefix.String())) == 0 {
+		response = true
+	}
+
+	return response
+}
+
 func (i *ReconcileIteration) deleteStaleDependencies() (errors []error) {
 	i.Log.Debug("Deleting stale dependencies")
 
@@ -210,9 +228,9 @@ func (i *ReconcileIteration) deleteStaleDependencies() (errors []error) {
 	if err != nil {
 		errors = append(errors, err)
 	} else {
-		i.Log.Debug("AllConnectors", "connectors", connectors)
 		for _, connector := range connectors.Items {
-			if !utils.ContainsString(connectorsToKeep, connector.GetName()) {
+			i.Log.Debug("AllConnectors", "connector", connector.GetName())
+			if !utils.ContainsString(connectorsToKeep, connector.GetName()) && i.isXJoinResource(connector.GetName()) {
 				i.Log.Info("Removing stale connector", "connector", connector.GetName())
 				if err = i.Kafka.DeleteConnector(connector.GetName()); err != nil {
 					errors = append(errors, err)
@@ -228,7 +246,7 @@ func (i *ReconcileIteration) deleteStaleDependencies() (errors []error) {
 	} else {
 		i.Log.Debug("AllESPipelines", "pipelines", esPipelines)
 		for _, esPipeline := range esPipelines {
-			if !utils.ContainsString(esPipelinesToKeep, esPipeline) {
+			if !utils.ContainsString(esPipelinesToKeep, esPipeline) && i.isXJoinResource(esPipeline) {
 				i.Log.Info("Removing stale es pipeline", "esPipeline", esPipeline)
 				if err = i.ESClient.DeleteESPipelineByFullName(esPipeline); err != nil {
 					errors = append(errors, err)
@@ -244,7 +262,7 @@ func (i *ReconcileIteration) deleteStaleDependencies() (errors []error) {
 	} else {
 		i.Log.Debug("AllIndices", "indices", indices)
 		for _, index := range indices {
-			if !utils.ContainsString(esIndicesToKeep, index) {
+			if !utils.ContainsString(esIndicesToKeep, index) && i.isXJoinResource(index) {
 				i.Log.Info("Removing stale index", "index", index)
 				if err = i.ESClient.DeleteIndexByFullName(index); err != nil {
 					errors = append(errors, err)
@@ -260,7 +278,7 @@ func (i *ReconcileIteration) deleteStaleDependencies() (errors []error) {
 	} else {
 		i.Log.Debug("AllTopics", "topics", topics)
 		for _, topic := range topics {
-			if !utils.ContainsString(topicsToKeep, topic) {
+			if !utils.ContainsString(topicsToKeep, topic) && i.isXJoinResource(topic) {
 				i.Log.Info("Removing stale topic", "topic", topic)
 				if err = i.Kafka.DeleteTopic(topic); err != nil {
 					errors = append(errors, err)
@@ -276,7 +294,7 @@ func (i *ReconcileIteration) deleteStaleDependencies() (errors []error) {
 	} else {
 		i.Log.Debug("AllReplicationSlots", "slots", slots)
 		for _, slot := range slots {
-			if !utils.ContainsString(replicationSlotsToKeep, slot) {
+			if !utils.ContainsString(replicationSlotsToKeep, slot) && i.isXJoinResource(slot) {
 				i.Log.Info("Removing stale replication slot", "slot", slot)
 				if err = i.InventoryDb.RemoveReplicationSlot(slot); err != nil {
 					errors = append(errors, err)
